Fixes action sets installed when multiple matches are used
For example previously installing a link for VLAN 100 between port 1 and 2:
1.100 <-> 2.100
Would be broken once a VLAN for port 1 and 3 on VLAN 200 was installed:
1.200 <-> 3.200

This is because a group was used per port. Therefore 1.200 replaces 1.100 and
in general confuses things.

Fixes bug when deleting flows, these never had the correct match or priority
so they would not delete. However, this continued to work as deleting the
group (per OF spec) will delete them. It is nice to be explict, in case a
switch implements this incorrectly.

General refactoring moving group, priority and match to functions in the Port
class, as these are used in a few places.